Fixing everything!!!

+++ src/MP3Track.cpp
@@ -36,7 +36,7 @@ void MP3Track::analyze_beatgrid()
     std::cout << "[MP3Track::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
-    double estimated_beats = (duration_seconds / 60.0) * bpm; // maybe change to long long type
+    long long estimated_beats = (duration_seconds / 60.0) * bpm; // maybe change to long long type
     double precision_factor = bitrate / 320.0;
     std::cout << "  → Estimated beats: " << estimated_beats << "  → Compression precision factor: " << precision_factor<< "\n";
 }
